# Use Seek() to create new wal segment