When doing a wal segment rollover, if we allocate and write & sync the entire segment, we are imposing a latency bump while the whole segments is synced to disk. 

Instead we can just seek to the size write 1 byte and we're ready to mmap the whole segment.

Before: 

```
Jul 10 09:02:25.350762 INF Stats - Total ops: 100455.4 ops/s - Failed ops:    0.0 ops/s
			Write ops 100455.4 w/s  Latency ms: 50%   6.5 - 95%   8.1 - 99%  93.3 - 99.9%  93.3 - max   93.3
			Read  ops    0.0 r/s  Latency ms: 50%   0.0 - 95%   0.0 - 99%   0.0 - 99.9%   0.0 - max    0.0
```

After:

```
Jul 10 09:01:07.853821 INF Stats - Total ops: 109892.6 ops/s - Failed ops:    0.0 ops/s
			Write ops 109892.6 w/s  Latency ms: 50%   6.6 - 95%  13.6 - 99%  40.2 - 99.9%  40.2 - max   40.2
			Read  ops    0.0 r/s  Latency ms: 50%   0.0 - 95%   0.0 - 99%   0.0 - 99.9%   0.0 - max    0.0
```

Note: forcing to write the entire segment could make sense because all the block are pre-allocated, though we should be doing it in more complex way: eg: pre-allocating segments in background and renaming the files when doing the rollover.